+++ rand/rand.h
@@ -5,7 +5,7 @@
 #define RAND_MAX 32767
 
 // Sets a custom seed value
-void srand(unsigned int seed);
+void srand(size_t seed);
 
 // Returns a psudorandom number between 0
 double rand(void);
